feature(Auth): get accessToken in gqlLink
