+++ src/apollo.client.ts
@@ -3,7 +3,7 @@ import { GraphQLWsLink } from "@apollo/client/link/subscriptions";
 import { getMainDefinition } from "@apollo/client/utilities";
 import { createClient } from "graphql-ws";
 import { setContext } from "@apollo/client/link/context";
-import { ACCESS_TOKEN_KEY } from "./local-storage";
+import { ACCESS_TOKEN_KEY, IDENTITY_TOKEN_KEY } from "./local-storage";
 
 export const createApolloClient = (httpUrl: string) => {
   const origin = new URL(httpUrl).origin;
@@ -14,8 +14,14 @@ export const createApolloClient = (httpUrl: string) => {
     uri: graphqlUrl,
   });
 
-  const authLink = setContext((_, { headers }) => {
-    const accessToken = localStorage.getItem(ACCESS_TOKEN_KEY);
+  const authLink = setContext(async (_, { headers }) => {
+    var accessToken = localStorage.getItem(ACCESS_TOKEN_KEY);
+    if (!accessToken) {
+      const identityToken = localStorage.getItem(IDENTITY_TOKEN_KEY);
+      if (identityToken) {
+        accessToken = await getAccessToken(identityToken, httpUrl);
+      }
+    }
     return {
       headers: {
         ...headers,
@@ -46,4 +52,29 @@ export const createApolloClient = (httpUrl: string) => {
     link: splitLink,
     cache: new InMemoryCache(),
   });
+};
+
+const getAccessToken = async (
+  identityToken: string,
+  webhookStoreUrl: string
+): Promise<string> => {
+  const webhookStoreHostname = new URL(webhookStoreUrl).hostname;
+  const accessTokenRequest = await fetch(
+    `${import.meta.env.VITE_AUTH_TENANT_URL}/webhook-store-auth/access-token`,
+    {
+      headers: {
+        Authorization: `Bearer ${identityToken}`,
+        "Content-Type": "application/json",
+      },
+      method: "POST",
+      body: JSON.stringify({
+        webhookStoreUrl: webhookStoreHostname,
+      }),
+    }
+  );
+  const json = await accessTokenRequest.json();
+  const accessToken = json.accessToken;
+  localStorage.setItem(ACCESS_TOKEN_KEY, accessToken);
+
+  return accessToken;
 };
